PSygrid force int input
force the user to input an integer. This includes numpy versions of int.

If this is not forces, the `initial_values` and `final_values` can be loaded with a 1 value array/list, but the `history1`, etc are not loaded with that input. As a result, `None` is returned for those and it looks like the history is empty.