PSygrid force int input (#440)
* PSygrid force int input

force the user to input an integer. This includes numpy versions of int.

If this is not forces, the `initial_values` and `final_values` can be loaded with a 1 value array/list, but the `history1`, etc are not loaded with that input. As a result, `None` is returned for those and it looks like the history is empty.

* add extra check for `contains`

raises an error in index in `__contains__` and `__getitem__` is not an integer

* Change error type of IndexError

Preferred way to raise error:
"""if key is a value outside the set of indexes for the sequence (after any special interpretation of negative values), IndexError should be raised. For mapping types, if key is missing (not in the container), KeyError should be raised."""

https://docs.python.org/3/reference/datamodel.html#object.__getitem__

* change to 'index'

* edited docstring to match arg name

---------

Co-authored-by: Elizabeth Teng <41969755+elizabethteng@users.noreply.github.com>